# storage/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LouvorJA API",
        "description": "API do LouvorJA - Gerenciador de Hinários e Músicas para a Igreja Adventista do Sétimo Dia. Fornece endpoints públicos para consulta de músicas, letras, arquivos e álbuns, além de endpoints administrativos protegidos por autenticação JWT.",
        "contact": {
            "name": "LouvorJA",
            "email": "contato@louvorja.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.louvorja.com.br",
            "description": "Produção"
        },
        {
            "url": "http://localhost:8000",
            "description": "Desenvolvimento"
        }
    ],
    "paths": {
        "/{lang}/albums": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Listar álbuns (público)",
                "description": "Retorna lista paginada de álbuns para o idioma informado",
                "operationId": "4c40849d962c5ca87686cf2d59db7418",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "Código do idioma",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "pt"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Busca textual",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de álbuns",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/admin/albums": {
            "get": {
                "tags": [
                    "Admin - Álbuns"
                ],
                "summary": "Listar álbuns",
                "description": "Retorna lista paginada de álbuns, com suporte a filtros por idioma e busca textual",
                "operationId": "4c96441c818261199ac28dea8f7c2c4c",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Idioma",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "pt"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Busca textual",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de álbuns",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Álbuns"
                ],
                "summary": "Criar álbum",
                "description": "Cria um novo(a) álbum. Requer autenticação admin.",
                "operationId": "63ccfd3316d5bb58336e507cef6651c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Álbum criado(a) com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados de validação inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/{lang}/albums/category/{slug}": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Álbuns por slug de categoria (público)",
                "description": "Retorna lista paginada de álbuns pertencentes a uma categoria identificada por slug. Slug inexistente retorna lista vazia.",
                "operationId": "a17003ca10060f55378565425c7ccc38",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "Código do idioma",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "pt"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug da categoria",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de álbuns",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/{lang}/albums/{id}": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Buscar álbum por ID (público)",
                "description": "Retorna os dados detalhados de um álbum para o idioma informado",
                "operationId": "5455f57dcbbbf4cb96914dc54bfe85f5",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "Código do idioma",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "pt"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do álbum",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do álbum",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Álbum não encontrado"
                    }
                },
                "security": []
            }
        },
        "/{lang}/album/{id}": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Buscar álbum por ID (alias singular)",
                "description": "Alias para /{lang}/albums/{id}",
                "operationId": "9a81d11285a42e09f28da11924d93207",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "Código do idioma",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "pt"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do álbum",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do álbum",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Álbum não encontrado"
                    }
                },
                "security": []
            }
        },
...
        "/{lang}/categories/{id}/albums": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Álbuns de uma categoria (público)",
                "description": "Retorna lista paginada de álbuns pertencentes a uma categoria específica",
                "operationId": "b9918ec07ef48c1c50258e0521b0b693",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "Código do idioma",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "pt"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de álbuns",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação JWT via header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}